Removed folder in the wrong place
