--- 7/main.py
@@ -1,4 +1,0 @@
-from game import DungeonGame
-
-current_game = DungeonGame()
-current_game.game_loop()
